+++ unpack-MTK.pl
@@ -23,6 +23,7 @@
 #   - added support for new platforms - MT6595 (thanks to carliv@XDA) (29-12-2014)
 #   - code cleanup and revised information output for boot and recovery images (29-12-2014)
 #   - make scripts more future-proof by supporting even more args (30-12-2014)
+#   - added new cmdline option for debugging purposes (06-01-2015)
 #
 
 use v5.14;
@@ -34,19 +35,22 @@ use Compress::Zlib;
 use Term::ANSIColor;
 use Scalar::Util qw(looks_like_number);
 use FindBin qw($Bin);
+use Text::Wrap;
 
-my $version = "MTK-Tools by Bruno Martins\nMTK unpack script (last update: 30-12-2014)\n";
-my $usageMain = "unpack-MTK.pl <infile> [COMMAND ...]\n  Unpacks MediaTek boot, recovery or logo images\n\nOptional COMMANDs are:\n\n";
-my $usageBootOpts =  "  -info_only\n    Display boot or recovery image information only\n     (useful to check image information without unpacking)\n\n  -kernel_only\n    Extract kernel only from boot or recovery image\n\n  -ramdisk_only\n    Extract ramdisk only from boot or recovery image\n\n";
-my $usageLogoOpts =  "  -force_logo_res <width> <height>\n    Forces logo image file to be unpacked by specifying image resolution,\n    which must be entered in pixels\n     (only useful when no zlib compressed images are found)\n\n  -invert_logo_res\n    Invert image resolution (width <-> height)\n     (may be useful when extracted images appear to be broken)\n\n";
+my $version = "MTK-Tools by Bruno Martins\nMTK unpack script (last update: 06-01-2015)\n";
+my $usageMain = "unpack-MTK.pl <infile> [COMMAND ...]\n  Unpacks MediaTek boot, recovery or logo images\n\n";
+my $usageBootOpts =  "Optional COMMANDs for boot or recovery images are:\n\n  -info_only\n    Display file information only\n     (useful to check image information without unpacking)\n\n  -kernel_only [--debug]\n    Extract kernel only\n\n  -ramdisk_only [--debug]\n    Extract ramdisk only\n\n" . wrap("    ","     ","(optional argument '--debug' can additionally be used to provide useful information for debugging purposes, even while unpacking both kernel and ramdisk)") . "\n\n";
+my $usageLogoOpts =  "Optional COMMANDs for logo images are:\n\n  -force_logo_res <width> <height>\n" . wrap("    ","     ","Forces file to be unpacked by specifying image resolution (in pixels)\n(only useful when no zlib compressed images are found)") . "\n\n  -invert_logo_res\n" . wrap("    ","     ","Invert image resolution (width <-> height)\n(may be useful when extracted images appear to be broken)") ."\n\n";
 my $usage = $usageMain . $usageBootOpts . $usageLogoOpts;
 
 print colored ("$version", 'bold blue') . "\n";
 die "Usage: $usage" unless $ARGV[0];
 
 if ($ARGV[1]) {
-	if ($ARGV[1] eq "-info_only" || $ARGV[1] eq "-kernel_only" || $ARGV[1] eq "-ramdisk_only" || $ARGV[1] eq "-invert_logo_res") {
+	if ($ARGV[1] eq "-info_only" || $ARGV[1] eq "-invert_logo_res" || $ARGV[1] eq "--debug") {
 		die "Usage: $usage" unless !$ARGV[2];
+	} elsif ($ARGV[1] eq "-kernel_only" || $ARGV[1] eq "-ramdisk_only") {
+		die "Usage: $usage" unless (!$ARGV[2] || $ARGV[2] eq "--debug" && !$ARGV[3]);
 	} elsif ($ARGV[1] eq "-force_logo_res") {
 		die "Usage: $usage" unless (looks_like_number($ARGV[2]) && looks_like_number($ARGV[3]) && !$ARGV[4]);
 	} else {
@@ -69,8 +73,8 @@ if ((substr($input, 0, 4) eq "\x88\x16\x88\x58") & (substr($input, 8, 4) eq "LOG
 	# if the input file contains the logo signature, try to unpack it
 	print "Valid logo signature found...\n";
 	if ($ARGV[1]) {
-		die_msg("$ARGV[1] switch can't be used with logo images!")
-			if ($ARGV[1] ne "-force_logo_res" && $ARGV[1] ne "-invert_logo_res");
+		die_msg("argument '$ARGV[1]' can't be used with logo images!")
+			unless ($ARGV[1] eq "-force_logo_res" || $ARGV[1] eq "-invert_logo_res");
 		$ARGV[1] =~ s/-//;
 		unpack_logo($input, $ARGV[1]);
 	} else {
@@ -80,20 +84,21 @@ if ((substr($input, 0, 4) eq "\x88\x16\x88\x58") & (substr($input, 8, 4) eq "LOG
 	# else, if a valid Android signature is found, try to unpack boot or recovery image
 	print "Valid Android signature found...\n";
 	if ($ARGV[1]) {
-		die_msg("$ARGV[1] switch can't be used with boot or recovery images!")
-			if ($ARGV[1] ne "-info_only" && $ARGV[1] ne "-kernel_only" && $ARGV[1] ne "-ramdisk_only");
+		die_msg("argument '$ARGV[1]' can't be used with boot or recovery images!")
+			unless ($ARGV[1] eq "-info_only" || $ARGV[1] eq "-kernel_only" || $ARGV[1] eq "-ramdisk_only" ||
+				$ARGV[1] eq "--debug");
 		$ARGV[1] =~ s/-//;
 		$ARGV[1] =~ s/_only//;
-		unpack_boot($input, $ARGV[1]);
+		unpack_boot($input, $ARGV[2] ? $ARGV[1] : "kernel and ramdisk", $ARGV[2] ? $ARGV[2] : $ARGV[1]);
 	} else {
-		unpack_boot($input, "kernel and ramdisk");
+		unpack_boot($input, "kernel and ramdisk", $ARGV[2]);
 	}
 } else {
 	die_msg("the input file does not appear to be supported or valid!");
 }
 
 sub unpack_boot {
-	my ($bootimg, $extract) = @_;
+	my ($bootimg, $extract, $mode) = @_;
 	my ($bootMagic, $kernelSize, $kernelLoadAddr, $ram1Size, $ram1LoadAddr, $ram2Size, $ram2LoadAddr, $tagsAddr, $pageSize, $unused1, $unused2, $bootName, $cmdLine, $id) = unpack('a8 L L L L L L L L L L a16 a512 a20', $bootimg);
 	my $magicAddr = 0x00000000;
 	my $baseAddr = $kernelLoadAddr - 0x00008000;
@@ -101,34 +106,37 @@ sub unpack_boot {
 	my $ram1Offset = $ram1LoadAddr - $baseAddr;
 	my $ram2Offset = $ram2LoadAddr - $baseAddr;
 	my $tagsOffset = $tagsAddr - $baseAddr;
+	my $debug_mode = ($mode =~ /debug/ ? 1 : 0);
 	my $unpack_sucess = 0;
 
 	# remove trailing zeros from board and cmdline
 	$bootName =~ s/\x00+$//;
 	$cmdLine =~ s/\x00+$//;
 
-	# print input file information
-	print colored ("\nInput file information:\n", 'cyan') . "\n";
-	print colored (" Header:\n", 'cyan') . "\n";
-	printf ("  Boot magic:\t\t\t%s\n", $bootMagic);
-	printf ("  Kernel size (bytes):\t\t%d\t\t(0x%.8x)\n", $kernelSize, $kernelSize);
-	printf ("  Kernel load address:\t\t0x%.8x\n\n", $kernelLoadAddr);
-	printf ("  Ramdisk size (bytes):\t\t%d\t\t(0x%.8x)\n", $ram1Size, $ram1Size);
-	printf ("  Ramdisk load address:\t\t0x%.8x\n", $ram1LoadAddr);
-	printf ("  Second stage size (bytes):\t%d\t\t(0x%.8x)\n", $ram2Size, $ram2Size);
-	printf ("  Second stage load address:\t0x%.8x\n\n", $ram2LoadAddr);
-	printf ("  Tags address:\t\t\t0x%.8x\n", $tagsAddr);
-	printf ("  Page size (bytes):\t\t%d\t\t(0x%.8x)\n", $pageSize, $pageSize);
-	printf ("  ASCIIZ product name:\t\t'%s'\n", $bootName);
-	printf ("  Command line:\t\t\t'%s'\n", $cmdLine);
-	printf ("  ID:\t\t\t\t%s\n\n", unpack('H*', $id));
-	print colored (" Other:\n", 'cyan') . "\n";
-	printf ("  Boot magic offset:\t\t0x%.8x\n", $magicAddr);
-	printf ("  Base address:\t\t\t0x%.8x\n\n", $baseAddr);
-	printf ("  Kernel offset:\t\t0x%.8x\n", $kernelOffset);
-	printf ("  Ramdisk offset:\t\t0x%.8x\n", $ram1Offset);
-	printf ("  Second stage offset:\t\t0x%.8x\n", $ram2Offset);
-	printf ("  Tags offset:\t\t\t0x%.8x\n\n", $tagsOffset);
+	# print input file information (only in normal mode)
+	if (!$debug_mode) {
+		print colored ("\nInput file information:\n", 'cyan') . "\n";
+		print colored (" Header:\n", 'cyan') . "\n";
+		printf ("  Boot magic:\t\t\t%s\n", $bootMagic);
+		printf ("  Kernel size (bytes):\t\t%d\t\t(0x%.8x)\n", $kernelSize, $kernelSize);
+		printf ("  Kernel load address:\t\t0x%.8x\n\n", $kernelLoadAddr);
+		printf ("  Ramdisk size (bytes):\t\t%d\t\t(0x%.8x)\n", $ram1Size, $ram1Size);
+		printf ("  Ramdisk load address:\t\t0x%.8x\n", $ram1LoadAddr);
+		printf ("  Second stage size (bytes):\t%d\t\t(0x%.8x)\n", $ram2Size, $ram2Size);
+		printf ("  Second stage load address:\t0x%.8x\n\n", $ram2LoadAddr);
+		printf ("  Tags address:\t\t\t0x%.8x\n", $tagsAddr);
+		printf ("  Page size (bytes):\t\t%d\t\t(0x%.8x)\n", $pageSize, $pageSize);
+		printf ("  ASCIIZ product name:\t\t'%s'\n", $bootName);
+		printf ("  Command line:\t\t\t'%s'\n", $cmdLine);
+		printf ("  ID:\t\t\t\t%s\n\n", unpack('H*', $id));
+		print colored (" Other:\n", 'cyan') . "\n";
+		printf ("  Boot magic offset:\t\t0x%.8x\n", $magicAddr);
+		printf ("  Base address:\t\t\t0x%.8x\n\n", $baseAddr);
+		printf ("  Kernel offset:\t\t0x%.8x\n", $kernelOffset);
+		printf ("  Ramdisk offset:\t\t0x%.8x\n", $ram1Offset);
+		printf ("  Second stage offset:\t\t0x%.8x\n", $ram2Offset);
+		printf ("  Tags offset:\t\t\t0x%.8x\n", $tagsOffset);
+	}
 
 	if ($extract eq "info") {
 		die colored ("Successfully displayed input file information.", 'green') . "\n";
@@ -139,7 +147,7 @@ sub unpack_boot {
 		or die_msg("couldn't create file '$inputFilename-args.txt'!");
 	printf ARGSFILE ("--base %#.8x\n--pagesize %d\n--kernel_offset %#.8x\n--ramdisk_offset %#.8x\n--second_offset %#.8x\n--tags_offset %#.8x%s%s", $baseAddr, $pageSize, $kernelOffset, $ram1Offset, $ram2Offset, $tagsOffset, $bootName eq "" ? "" : "\n--board $bootName", $cmdLine eq "" ? "" : "\n--cmdline $cmdLine") or die;
 	close (ARGSFILE);
-	print "Extra arguments written to '$inputFilename-args.txt'\n";
+	print "\nExtra arguments written to '$inputFilename-args.txt'\n";
 
 	if ($extract =~ /kernel/) {
 		my $kernel = substr($bootimg, $pageSize, $kernelSize);
@@ -185,9 +193,13 @@ sub unpack_boot {
 			die_msg("'$tool' binary not found! Double check your environment setup.")
 				if system ("command -v $tool >/dev/null 2>&1");
 		}
+		if ($debug_mode) {
+			print colored ("\nRamdisk unpack command:", 'yellow') . "\n";
+			print "'gzip -d -c ../$inputFilename-ramdisk.cpio.gz | cpio -i'\n\n";
+		}
 		print "Ramdisk size: ";
 		system ("gzip -d -c ../$inputFilename-ramdisk.cpio.gz | cpio -i");
-		system ("rm ../$inputFilename-ramdisk.cpio.gz");
+		system ("rm ../$inputFilename-ramdisk.cpio.gz") unless ($debug_mode);
 
 		print "Extracted ramdisk contents to directory '$inputFilename-ramdisk'\n";
 		$unpack_sucess = 1;
@@ -255,7 +267,7 @@ sub unpack_logo {
 
 	if (!$num_blocks) {
 		die_msg("no zlib packed rgb565 images were found inside logo file!" . 
-			"\nDouble check script usage and try using -force_logo_res switch.") unless ($switch eq "force_logo_res");
+			"\nDouble check script usage and try using '-force_logo_res' argument.") unless ($switch eq "force_logo_res");
 
 		# if no compressed files are found, try to unpack logo based on specified image resolution
 		my $image_file_size = ($ARGV[2] * $ARGV[3] * 2);
@@ -365,6 +377,6 @@ sub rgb565_to_png {
 }
 
 sub die_msg {
-	die colored ("\nError: $_[0]", 'red') . "\n";
+	die colored ("\n" . wrap("","","Error: $_[0]"), 'red') . "\n";
 }
